A bunch of stuff!
I removed the dependency I added today as we don't seem to need it and I'm not really sure what it's for! 

I also got the slideshow to work - I think this will be a better option as we can't seem to add buttons inside of lightbox (I will be adding temp buttons to the puppies page in a bit). I did change the links in my local seeds though, because a lot of the photos weren't the right size for this particular feature...I'll show you guys what I have in the morning and if you like it we can change the db seeds to include the new ones. Thanks!!